Add a new challenge: the Levenshtein distance
The Levenshtein distance is a metric which I have long been a fan
of for how simple and fun it is to (often recursively) calculate.
We teach this to first year computer science students  as well. Of
course, they are discouraged from writing unreadably tiny code.

The code which I have written here is absolutely not tiny and is
just meant as a challenge for others to minify it.